codespace_gemini_browser_demo/requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
google-genai==1.56.0
